Bug dans ft_strrchr size_t en int

+++ Libft/ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: rlarbi <rlarbi@student.42nice.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 17:25:48 by rlarbi            #+#    #+#             */
-/*   Updated: 2024/11/14 14:39:15 by rlarbi           ###   ########.fr       */
+/*   Updated: 2024/11/21 13:07:46 by rlarbi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+	int	i;
 
 	i = ft_strlen(s);
 	while (i >= 0)
